feat: Permission setting, View count
 - test removed

+++ src/main/java/com/caaasperr/Alcoholic/domain/step/service/StepService.java
@@ -1,9 +1,11 @@
 package com.caaasperr.Alcoholic.domain.step.service;
 
+import com.caaasperr.Alcoholic._common.annotation.CheckCocktailOwner;
 import com.caaasperr.Alcoholic._common.exception.CustomException;
 import com.caaasperr.Alcoholic._common.exception.ErrorCode;
 import com.caaasperr.Alcoholic._common.exception.IllegalArgumentException;
 import com.caaasperr.Alcoholic._common.image.ImageHandler;
+import com.caaasperr.Alcoholic.domain.cocktail.model.Cocktail;
 import com.caaasperr.Alcoholic.domain.cocktail.repository.CocktailRepository;
 import com.caaasperr.Alcoholic.domain.step.dto.CocktailStep;
 import com.caaasperr.Alcoholic.domain.step.dto.CreateStepRequest;
@@ -11,6 +13,7 @@ import com.caaasperr.Alcoholic.domain.step.dto.UpdateStepRequest;
 import com.caaasperr.Alcoholic.domain.step.model.Step;
 import com.caaasperr.Alcoholic.domain.step.repository.StepRepository;
 import jakarta.transaction.Transactional;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -33,11 +36,16 @@ public class StepService {
     }
 
     @Transactional
-    public void createStep(CreateStepRequest request) throws IOException {
+    public void createStep(CreateStepRequest request, Authentication authentication) throws IOException {
+        Cocktail cocktail = cocktailRepository.findById(request.cocktail_id()).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_COCKTAIL));
+        if (!cocktail.getUser().getUsername().equals(authentication.getName())) {
+            throw new CustomException(ErrorCode.ACCESS_DENIED);
+        }
+
         String imagePath = request.image() != null ? imageHandler.saveImage(request.image()) : null;
         stepRepository.save(
                 request.toStep(
-                        cocktailRepository.findById(request.cocktail_id()).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_COCKTAIL)),
+                        cocktail,
                         imagePath
                 )
         );
@@ -48,8 +56,12 @@ public class StepService {
     }
 
     @Transactional
-    public void updateStep(Long id, UpdateStepRequest request) throws IOException {
+    public void updateStep(Long id, UpdateStepRequest request, Authentication authentication) throws IOException {
         Step step = stepRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_COCKTAIL));
+
+        if (!step.getCocktail().getUser().getUsername().equals(authentication.getName())) {
+            throw new CustomException(ErrorCode.ACCESS_DENIED);
+        }
 
         if (request.content() != null && !step.getContent().equals(request.content())) {
             step.updateContent(request.content());
@@ -69,8 +81,9 @@ public class StepService {
         }
     }
 
+    @CheckCocktailOwner
     @Transactional
-    public void reorderSteps(Long cocktailId, List<Long> orderedStepIds) {
+    public void reorderSteps(Long cocktailId, List<Long> orderedStepIds, Authentication authentication) {
         List<Step> steps = stepRepository.findByCocktail_id(cocktailId);
 
         if (steps.size() != orderedStepIds.size()) {
@@ -92,7 +105,13 @@ public class StepService {
         }
     }
 
-    public void deleteStep(Long id) {
+    public void deleteStep(Long id, Authentication authentication) {
+        Step step = stepRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_STEP));
+
+        if (!step.getCocktail().getUser().getUsername().equals(authentication.getName())) {
+            throw new CustomException(ErrorCode.ACCESS_DENIED);
+        }
+
         stepRepository.deleteById(id);
     }
 }
